pyproject.toml
[project]
name = "printerbridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Théo Lévesque", email = "theo.levesque.024@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pyusb>=1.3.1",
]

[project.scripts]
printerbridge = "printerbridge:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
